Rechecked version of kmp_C4-179_pycharm_support, but with latest poetry.
