+++ pyproject.toml
@@ -36,8 +36,9 @@ classifiers = [
 
 [tool.poetry.dependencies]
 python = ">=3.6.1,<3.7"
-boto3 = "^1.10.50"
-botocore = "^1.13.50"
+awscli = ">=1.18.219"
+boto3 = "^1.16.59"
+botocore = "^1.19.59"
 certifi = ">=2020.11.8"
 chardet = "3.0.4"
 colorama = "0.3.3"
@@ -49,82 +50,82 @@ elasticsearch = "6.8.1"
 elasticsearch-dsl = "^6.4.0"
 execnet = "1.4.1"
 future = "^0.15.2"
-humanfriendly = "1.44.5"
+humanfriendly = "^1.44.9"
 hupper = "1.5"
-idna = ">=2.7,<3"  # was "2.7"
+idna = ">=2.7,<3"
 isodate = "0.5.4"
 jmespath = "0.9.0"
-jsonschema-serialize-fork = "2.1.1"
+jsonschema_serialize_fork = "2.1.1"
 keepalive = "0.5"
 loremipsum = "1.0.5"
-netaddr = ">=0.7.18,<1"
+netaddr = ">=0.8.0,<1"
 negspy = "0.2.24"
 numpy = "1.19.1" # negspy needs it
-passlib = "1.6.5"
+passlib = "^1.7.4"
 PasteDeploy = "1.5.2"
 pbkdf2 = "1.3"
 Pillow = "^6.2.2"  # 6.x.x should work for CGAP -Will 11/18/2020
 plaster = "1.0"
 plaster-pastedeploy = "0.6"
-psutil = "^5.6.6"
-psycopg2-binary = "2.7.7"
-py = "^1.8.1"  # was "1.4.31" (used by pytest, not sure if elsewhere)
+psutil = "^5.8.0"
+psycopg2-binary = "^2.7.7"
+py = ">=1.8.1"  # was "1.4.31" (used by pytest, not sure if elsewhere)
 pyasn1 = "0.1.9"
 PyBrowserID = "^0.10.0"
 pycparser = "2.14"
 PyJWT = "1.5.3"
-pyparsing = "^2.4.7"  # was "2.1.1"
+pyparsing = "^2.4.7"
 pyramid = "1.10.4"
-pyramid-localroles = ">=0.1,<1"
-pyramid-multiauth = "0.8.0"
-pyramid-retry = "1.0"
-pyramid-tm = "2.2.1"
-pyramid-translogger = "^0.1"
-python-dateutil = "2.7.3"
+pyramid_localroles = ">=0.1,<1"
+pyramid_multiauth = ">=0.9.0,<1"
+pyramid-retry = "^1.0"
+pyramid-tm = "^2.4"
+pyramid_translogger = "^0.1"
+python-dateutil = "^2.8.1"
 # python-magic is presently pinned to 0.4.15 in lockstep with dcicsnovault's requirements. See explanation there.
 python-magic = "0.4.15"
 pytz = ">=2020.4"
 rdflib = "^4.2.2"
-rdflib-jsonld = "0.3"
+rdflib-jsonld = ">=0.5.0,<1.0.0"
 requests = "^2.23.0"
-rfc3987 = "^1.3.6"
+rfc3987 = "^1.3.8"
 rsa = "3.3"
-rutter = ">=0.2,<1"
-s3transfer = "^0.2.0"
+rutter = ">=0.3,<1"
+s3transfer = ">=0.3.0,<0.4.0"
 simplejson = "^3.17.0"
-SPARQLWrapper = "1.7.6"
+SPARQLWrapper = "^1.7.6"
 SQLAlchemy = "1.3.16"  # Pinned because >=1.3.17 is broken for us (circular constraints prevent deletes)
-structlog = ">=18.1.0,<20"
-submit4dn = "0.9.7"
-subprocess-middleware = ">=0.3,<1"
+structlog = ">=19.2.0,<20"
+#submit4dn = "0.9.7"
+subprocess_middleware = ">=0.3,<1"
 # Useful for picking apart pyproject.toml
 toml = ">=0.10.1,<1"
 tqdm = "^4.51.0"
-transaction = "2.4.0"
+transaction = "^2.4.0"
 translationstring = "1.3"
 uptime = ">=3.0.1,<4"
 urllib3 = "^1.24.3"
 venusian = "^1.2.0"
 waitress = "1.2.0"  # 1.4.3 had lots of problems, so pin this -kmp 18-May-2020
-WebOb = "1.8.5"
-WebTest = "^2.0.21"
+WebOb = "^1.8.6"
+WebTest = "^2.0.35"
 wheel = "0.29.0"
 WSGIProxy2 = "0.4.2"
-xlrd = "^1.0.0"
+xlrd = "^1.2.0"
 xlwt = "1.2.0"
-"zope.deprecation" = "4.4.0"
-"zope.interface" = "4.6.0"
+"zope.deprecation" = "^4.4.0"
+"zope.interface" = "^4.7.2"
 "zope.sqlalchemy" = "1.3"
 sentry-sdk = "^0.16.5"
 
 [tool.poetry.dev-dependencies]
 # PyCharm says boto3-stubs contains useful type hints
 boto3-stubs = ">=1.16.15.0"
-coverage = ">=5.2"
+coverage = ">=5.3.1"
 codacy-coverage = ">=1.3.11"
-coveralls = ">=2.1.1"
-flake8 = "^3.7.8"
-flaky = "3.6.1"
+coveralls = ">=3.0.0"
+flake8 = ">=3.8.4"
+flaky = ">=3.6.1"
 # flask only for moto[server]
 flask = ">=1.1.1"
 # Here AND ELSEWHERE (related scripts), we pin use of moto 1.3.7.
@@ -138,14 +139,15 @@ flask = ">=1.1.1"
 # TODO: Investigate whether this bug is fixed. If it is, update the version.
 #       If not, get a reference to an open bug report here so we can check back periodically.
 moto = "1.3.7"
-pytest = "4.5.0"  # was "2.9.2"
-pytest-cov = "2.8.1"  # was "2.2.1"  (2.7.1 might also work)
-pytest-exact-fixtures = "0.3"  # was "0.1"
-pytest-instafail = "0.4.1"  # was "0.3.0"
-pytest-mock = "3.1.0"  # was "0.11.0" (1.10.4 might also work)
-pytest-timeout = "1.3.4"  # was "1.0.0"
-pytest-xdist = ">=1.14"  # was "1.14"
-"repoze.debug" = "1.1"  # was "1.0.2"
+pytest = "4.5.0"
+pytest-cov = ">=2.9.0"
+pytest-exact-fixtures = ">=0.3"
+pytest-instafail = ">=0.4.2"
+pytest-mock = ">=3.2.0"
+pytest-runner = ">=5.2"
+pytest-timeout = ">=1.4.2"
+pytest-xdist = ">=1.34"
+"repoze.debug" = ">=1.1"
 # responses = "0.10.14"
 
 # In pytest 6.0, we'll be able to use this instead of a separate pytest.ini configuration.
