+++ pytest.ini
@@ -1,9 +1,5 @@
 [pytest]
 addopts =
-    --pyargs encoded.tests deploy.tests
-# These could be plugins but src/encoded/tests/conftest.py hasn't traditionally listed them as such:
-#    -p encoded.tests.variant_fixtures
-#    -p encoded.tests.test_fixtures
    -p encoded.tests.datafixtures
    -p snovault.tests.serverfixtures
     --instafail
@@ -25,3 +21,6 @@ markers =
     unit: a proper unit test
     working: mark a test as working (deselect with '-m "not working"')
 norecursedirs = *env site-packages .cache .git .idea *.egg-info
+testpaths =
+    src/encoded
+    deploy
