+++ src/encoded/commands/purge_item_type.py
@@ -43,7 +43,7 @@ def purge_item_type_from_storage(app, item_types, prod=False):
 
     # purge uuids directly from PickStorage, ignoring status=deleted checks
     configure_dbsession(app)
-    uuids_to_purge = set(get_uuids_for_types(app.registry, item_types))
+    uuids_to_purge = get_uuids_for_types(app.registry, item_types)
     pstorage = app.registry[STORAGE]
     for uuid in uuids_to_purge:
         try:
