+++ src/encoded/renderers.py
@@ -22,14 +22,14 @@ from pyramid.traversal import split_path_info, _join_path_tuple
 from subprocess_middleware.worker import TransformWorker
 from urllib.parse import urlencode
 from webob.cookies import Cookie
-from .ingestion.common import content_type_allowed
+from .util import content_type_allowed
 
 
 log = logging.getLogger(__name__)
 
 
 def includeme(config):
-    '''
+    """
     Can get tween ordering by executing the following on command-line from root dir:
         `bin/ptween development.ini`
 
@@ -64,13 +64,15 @@ def includeme(config):
     This means that if handler(request) is called, then the downstream tweens are acted upon it,
     until response is returned. It's an ONION!
 
-    '''
+    """
 
     config.add_tween('.renderers.validate_request_tween_factory', under='snovault.stats.stats_tween_factory')
-    # DISABLED - .add_tween('.renderers.remove_expired_session_cookies_tween_factory', under='.renderers.validate_request_tween_factory')
+    # DISABLED - .add_tween('.renderers.remove_expired_session_cookies_tween_factory',
+    #                       under='.renderers.validate_request_tween_factory')
     config.add_tween('.renderers.render_page_html_tween_factory', under='.renderers.validate_request_tween_factory')
 
-    # The above tweens, when using response (= `handler(request)`) act on the _transformed_ response (containing HTML body).
+    # The above tweens, when using response (= `handler(request)`) act on the _transformed_ response
+    # (containing HTML body).
     # The below tweens run _before_ the JS rendering. Responses in these tweens have not been transformed to HTML yet.
     config.add_tween('.renderers.set_response_headers_tween_factory', under='.renderers.render_page_html_tween_factory')
 
@@ -127,7 +129,7 @@ def security_tween_factory(handler, registry):
         """
 
         expected_user = request.headers.get('X-If-Match-User')
-        if expected_user is not None: # Not sure when this is the case
+        if expected_user is not None:  # Not sure when this is the case
             if request.authenticated_userid != 'mailto.' + expected_user:
                 detail = 'X-If-Match-User does not match'
                 raise HTTPPreconditionFailed(detail)
@@ -142,15 +144,18 @@ def security_tween_factory(handler, registry):
                     raise HTTPUnauthorized(
                         title="No Access",
                         comment="Invalid Authorization header or Auth Challenge response.",
-                        headers={'WWW-Authenticate': "Bearer realm=\"{}\"; Basic realm=\"{}\"".format(request.domain, request.domain) }
+                        headers={
+                            'WWW-Authenticate': ("Bearer realm=\"{}\"; Basic realm=\"{}\""
+                                                 .format(request.domain, request.domain))
+                        }
                     )
-
 
         if hasattr(request, 'auth0_expired'):
             # Add some security-related headers on the up-swing
             response = handler(request)
             if request.auth0_expired:
-                #return response
+                # return response
+                #
                 # If have the attribute and it is true, then our session has expired.
                 # This is true for both AJAX requests (which have request.authorization) & browser page
                 # requests (which have cookie); both cases handled in authentication.py
@@ -161,9 +166,13 @@ def security_tween_factory(handler, registry):
                 # Especially for initial document requests by browser, but also desired for AJAX and other requests,
                 # unset jwtToken cookie so initial client-side React render has App(instance).state.session = false
                 # to be synced w/ server-side
-                response.set_cookie(name='jwtToken', value=None, max_age=0,path='/') # = Same as response.delete_cookie(..)
+                response.set_cookie(name='jwtToken',
+                                    value=None, max_age=0, path='/')  # = Same as response.delete_cookie(..)
                 response.status_code = 401
-                response.headers['WWW-Authenticate'] = "Bearer realm=\"{}\", title=\"Session Expired\"; Basic realm=\"{}\"".format(request.domain, request.domain)
+                response.headers['WWW-Authenticate'] = (
+                    "Bearer realm=\"{}\", title=\"Session Expired\"; Basic realm=\"{}\""
+                    .format(request.domain, request.domain)
+                )
             else:
                 # We have JWT and it's not expired. Add 'X-Request-JWT' & 'X-User-Info' header.
                 # For performance, only do it if should transform to HTML as is not needed on every request.
@@ -175,9 +184,12 @@ def security_tween_factory(handler, registry):
                             # This header is parsed in renderer.js, or, more accurately,
                             # by libs/react-middleware.js which is imported by server.js and compiled into
                             # renderer.js. Is used to get access to User Info on initial web page render.
-                            response.headers['X-Request-JWT'] = request.cookies.get('jwtToken','')
-                            user_info = request.user_info # Re-ified property set in authentication.py
-                            del user_info["id_token"] # Redundant - don't need this in SSR nor browser as get from X-Request-JWT.
+                            response.headers['X-Request-JWT'] = request.cookies.get('jwtToken', '')
+                            # TODO: Should user_info be copied before the del? If the user info is shared,
+                            #       we are modifying it for other uses. -kmp 24-Jan-2021
+                            user_info = request.user_info  # Re-ified property set in authentication.py
+                            # Redundant - don't need this in SSR nor browser as get from X-Request-JWT.
+                            del user_info["id_token"]
                             response.headers['X-User-Info'] = json.dumps(user_info)
                         else:
                             response.headers['X-Request-JWT'] = "null"
@@ -190,20 +202,20 @@ def security_tween_factory(handler, registry):
         # requests from Authorization header which acts like a CSRF token.
         # See authentication.py - get_jwt()
 
-        #token = request.headers.get('X-CSRF-Token')
-        #if token is not None:
-        #    # Avoid dirtying the session and adding a Set-Cookie header
-        #    # XXX Should consider if this is a good idea or not and timeouts
-        #    if token == dict.get(request.session, '_csrft_', None):
-        #        return handler(request)
-        #    raise CSRFTokenError('Incorrect CSRF token')
+        # token = request.headers.get('X-CSRF-Token')
+        # if token is not None:
+        #     # Avoid dirtying the session and adding a Set-Cookie header
+        #     # XXX Should consider if this is a good idea or not and timeouts
+        #     if token == dict.get(request.session, '_csrft_', None):
+        #         return handler(request)
+        #     raise CSRFTokenError('Incorrect CSRF token')
         # raise CSRFTokenError('Missing CSRF token')
 
     return security_tween
 
 
 def remove_expired_session_cookies_tween_factory(handler, registry):
-    '''
+    """
     CURRENTLY DISABLED
     Original purpose of this was to remove expired (session?) cookies.
     See: https://github.com/ENCODE-DCC/encoded/commit/75854803c99e5044a6a33aedb3a79d750481b6cd#diff-bc19a9793a1b3b4870cff50e7c7c9bd1R135
@@ -211,7 +223,7 @@ def remove_expired_session_cookies_tween_factory(handler, registry):
     We disable it for now via removing from tween chain as are using JWT tokens and handling
     their removal in security_tween_factory & authentication.py as well as client-side
     (upon "Logout" action). If needed for some reason, can re-enable.
-    '''
+    """  # noQA - not going to break the long URL line above
 
     ignore = {
         '/favicon.ico',
@@ -222,8 +234,8 @@ def remove_expired_session_cookies_tween_factory(handler, registry):
             return handler(request)
 
         session = request.session
-        #if session or session._cookie_name not in request.cookies:
-        #    return handler(request)
+        # if session or session._cookie_name not in request.cookies:
+        #     return handler(request)
 
         response = handler(request)
         # Below seems to be empty always; though we do have some in request.cookies
@@ -247,7 +259,7 @@ def remove_expired_session_cookies_tween_factory(handler, registry):
 
 
 def set_response_headers_tween_factory(handler, registry):
-    '''Add additional response headers here'''
+    """Add additional response headers here"""
     def set_response_headers_tween(request):
         response = handler(request)
         response.headers['X-Request-URL'] = request.url
@@ -318,7 +330,8 @@ def best_mime_type(request, mode=MIME_TYPE_TRIAGE_MODE):
     In the case we can't comply, we just use application/json whether or not that's what was asked for.
     """
     if mode == 'legacy':
-        # See: https://tedboy.github.io/flask/generated/generated/werkzeug.Accept.best_match.html#werkzeug-accept-best-match
+        # See:
+        # https://tedboy.github.io/flask/generated/generated/werkzeug.Accept.best_match.html#werkzeug-accept-best-match
         # Note that this is now deprecated, or will be. The message is oddly worded ("will be deprecated")
         # that presumably means "will be removed". Deprecation IS the warning of actual action, not the action itself.
         # "This is currently maintained for backward compatibility, and will be deprecated in the future.
@@ -345,13 +358,13 @@ def best_mime_type(request, mode=MIME_TYPE_TRIAGE_MODE):
 
 @lru_cache(maxsize=16)
 def should_transform(request, response):
-    '''
+    """
     Determines whether to transform the response from JSON->HTML/JS depending on type of response
     and what the request is looking for to be returned via e.g. request Accept, Authorization header.
     In case of no Accept header, attempts to guess.
 
     Memoized via `lru_cache`. Cache size is set to be 16 (> 1) in case sub-requests fired off during handling.
-    '''
+    """
     # We always return JSON in response to POST, PATCH, etc.
     if request.method not in ('GET', 'HEAD'):
         return False
@@ -360,7 +373,8 @@ def should_transform(request, response):
     if response.content_type != 'application/json':
         return False
 
-    # If we have a 'frame' that is not None or page, force JSON, since our UI doesn't handle all various forms of the data, just embedded/page.
+    # If we have a 'frame' that is not None or page, force JSON, since our UI doesn't handle all various
+    # forms of the data, just embedded/page.
     request_frame = request.params.get("frame", "page")
     if request_frame != "page":
         return False
@@ -374,16 +388,18 @@ def should_transform(request, response):
         if format == 'html':
             return True
         else:
-            raise HTTPNotAcceptable("Improper format URI parameter", comment="The format URI parameter should be set to either html or json.")
+            raise HTTPNotAcceptable("Improper format URI parameter",
+                                    comment="The format URI parameter should be set to either html or json.")
 
     # Web browsers send an Accept request header for initial (e.g. non-AJAX) page requests
     # which should contain 'text/html'
     # See: https://tedboy.github.io/flask/generated/generated/werkzeug.Accept.best_match.html#werkzeug-accept-best-match
     mime_type = best_mime_type(request)
-    format = mime_type.split('/', 1)[1] # Will be 1 of 'html', 'json', 'json-ld'
+    format = mime_type.split('/', 1)[1]  # Will be 1 of 'html', 'json', 'json-ld'
 
-    # N.B. ld+json (JSON-LD) is likely more unique case and might be sent by search engines (?) which can parse JSON-LDs.
-    # At some point we could maybe have it to be same as making an `@@object` or `?frame=object` request (?) esp if fill
+    # N.B. ld+json (JSON-LD) is likely more unique case and might be sent by search engines (?)
+    # which can parse JSON-LDs. At some point we could maybe have it to be same as
+    # making an `@@object` or `?frame=object` request (?) esp if fill
     # out @context response w/ schema(s) (or link to schema)
 
     if format == 'html':
@@ -410,7 +426,9 @@ def render_page_html_tween_factory(handler, registry):
 
     rss_limit = 256 * (1024 ** 2)  # MB
 
-    reload_process = True if registry.settings.get('reload_templates', False) else lambda proc: psutil.Process(proc.pid).memory_info().rss > rss_limit
+    reload_process = (True
+                      if registry.settings.get('reload_templates', False)
+                      else lambda proc: psutil.Process(proc.pid).memory_info().rss > rss_limit)
 
     # TransformWorker inits and manages a subprocess
     # it re-uses the subprocess so interestingly data in JS global variables
